requirements.txt
annotated-types==0.7.0
anyio==4.8.0
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
chatgpt_md_converter==0.2.0
click==8.1.8
colorama==0.4.6
Deprecated==1.2.18
distro==1.9.0
exceptiongroup==1.2.2
Flask==3.1.0
Flask-Admin==1.6.1
Flask-BasicAuth==0.2.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
MarkupSafe==3.0.2
netaddr==1.3.0
openai==1.60.0
psycopg2-binary==2.9.10
pydantic==2.10.4
pydantic_core==2.27.2
pyTelegramBotAPI==4.25.0
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.36
telebot==0.0.5
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
waitress==3.0.2
Werkzeug==3.1.3
wrapt==1.17.2
WTForms==3.1.0
yookassa==3.5.0
